refactor: group completionist components
